Update HTML/CSS to be more GitHub accurate (#9)
Add additional HTML around the rendered markdown to be more accurate to
GitHub's wrapping.

Add some CSS to make it more pleasant to look at.

Interestingly `horrorshow` has no way to render HTML in a pretty,
indented manner which made the file name in the header appear too close
to the book octicon. To fix it I had to put a goofy call to `format!` in
the markdown template to put a space in front of the file name.