Reload sort of working
Reload is working in that javascript currently refreshes the page every
three seconds. Because I implemented ETag and If-None-Match for all the
resources I had hoped it would be pretty invisible to the user but it's
still pretty bad.

I think I'll have to switch to having the javascript make a HEAD request
every few seconds using ETag and If-None-Match and only have it refresh
the page if it gets a 200 rather than a 304.