Add auto-refresh with SSE
Add auto-refresh with SSE. It's going to be a significant undertaking to
maintain the state of the file on the server. Right now the server
always sends the file hash and the file contents with every event. The
front end will check if the hash is different and update the content if
the hash changed.

I also set it to always render in offline mode since the SSE would be
calling the GitHub API every few seconds and get rate limited.